Update C# mops to use 32bit values like the rest of the tests.
git-svn-id: https://svn.parrot.org/parrot/trunk@1947 d31e2699-5ff4-0310-a27c-f18f2fbe73fe

+++ examples/mops/mops.cs
@@ -3,10 +3,10 @@ using System;
 
 class App {
 	public static int Main(String[] args) {
-		long i1, i5;
+		int i1, i5;
 		// int i2 = 0;
-		long i3 = 1;
-		long i4 = 100000000;
+		int i3 = 1;
+		int i4 = 100000000;
 		DateTime start, end;
 		double n1, n2;
 		Console.WriteLine("Iterations:    {0}", i4);
